+++ common/src/main/java/io/github/flemmli97/runecraftory/api/datapack/ShopItemProperties.java
@@ -1,0 +1,6 @@
+package io.github.flemmli97.runecraftory.api.datapack;
+
+import net.minecraft.world.item.ItemStack;
+
+public record ShopItemProperties(ItemStack stack, boolean needsSpecialUnlocking) {
+}