+++ common/src/main/java/io/github/flemmli97/runecraftory/common/lib/LibNBT.java
@@ -9,6 +9,7 @@ public class LibNBT {
     public static final String Level = "ItemLevel";
     public static final String MagnifyingGlass = "MagnifyingGlass";
     public static final String ScrapMetalPlus = "ScrapMetalPlus";
+    public static final String ObjectX = "ObjectX";
 
     public static final String WateringCanWater = "Water";